Increase content length. Add timeout on packs.

+++ src/main/java/com/gitlab/tixtix320/sonder/api/common/communication/Transfer.java
@@ -9,7 +9,7 @@ public interface Transfer {
 
 	ReadableByteChannel channel();
 
-	int getContentLength();
+	long getContentLength();
 
 	byte[] readAll() throws IOException;
 }
